requirements.txt
anthropic>=0.45.0
pydantic-settings>=2.0
click>=8.0
apscheduler>=3.10
Pillow>=10.0
aiosqlite>=0.19
pytest>=8.0
pytest-asyncio>=0.23
